Issues #5: Green and Refactoring:            -> Moved Unit Test Classes.
